seeder

+++ database/migrations/2024_06_20_194454_create_produks_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama');
             $table->string('deskripsi');
             $table->string('harga');
-            $table->integer('stok');
+            $table->float('stok');
             $table->string('satuan');
             $table->timestamps();
         });
