+++ database/migrations/2024_06_20_194504_create_bahans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('bahan', function (Blueprint $table) {
             $table->increments('id_bahan')->primary;
             $table->string('nama_bahan');
-            $table->integer('stok_bahan');
+            $table->float('stok_bahan');
             $table->string('satuan_bahan');
             $table->timestamps();
         });
